backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.12
swagger-ui-bundle==1.1.0

numpy==2.1.3
pdfplumber==0.11.5
python-docx==1.1.2
sentence-transformers==3.3.1

requests==2.32.3
